+++ lib/exdash/placeholder.ex
@@ -1,0 +1,5 @@
+defmodule Exdash.Placeholder do
+  @moduledoc ~S"""
+  Placeholder used for determining argument order
+  """
+end