+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Delegation functions in order to use `Exdash.<function>` instead of `Exdash.<module>.<function>`
 - Function.curry
 - Function.partial
+- Placeholder to Function.partial
 
 ### Change
 - Function.after_nth doctest